+++ aspnetcore/performance/caching/middleware/sample/.bowerrc
@@ -1,0 +1,3 @@
+{
+  "directory": "wwwroot/lib"
+}
