+++ aspnetcore/performance/caching/middleware/sample/Pages/Index.cshtml.cs
@@ -1,0 +1,11 @@
+﻿using Microsoft.AspNetCore.Mvc.RazorPages;
+
+namespace ResponseCachingMiddleware.Pages
+{
+    public class IndexModel : PageModel
+    {
+        public void OnGet()
+        {
+        }
+    }
+}
