+++ aspnetcore/performance/caching/middleware/sample/Pages/_ViewImports.cshtml
@@ -1,0 +1,3 @@
+@using ResponseCachingMiddleware
+@namespace ResponseCachingMiddleware.Pages
+@addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
