+++ aspnetcore/performance/caching/middleware/sample/appsettings.json
@@ -1,0 +1,8 @@
+﻿{
+  "Logging": {
+    "IncludeScopes": false,
+    "LogLevel": {
+      "Default": "Debug"
+    }
+  }
+}
