+++ aspnetcore/performance/caching/middleware/sample/wwwroot/css/site.css
@@ -1,0 +1,25 @@
+body {
+    padding-top: 50px;
+    padding-bottom: 20px;
+}
+
+h1 {
+    font-size: 24px;
+}
+
+h2 {
+    font-size: 20px;
+}
+
+h3 {
+    font-size:16px
+}
+
+.body-content {
+    padding-left: 15px;
+    padding-right: 15px;
+}
+
+.panel-body {
+    font-size: 16px;
+}
