+++ aspnetcore/performance/caching/middleware/sample/wwwroot/js/site.js
@@ -1,0 +1,1 @@
+﻿// Write your Javascript code.
